@@ -4,50 +4,55 @@
   "description": "GraphQL and React Prototype Application",
   "main": "index.js",
   "dependencies": {
-    "axios": "^0.19.2",
-    "classnames": "^2.2.6",
-    "moment": "^2.24.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "axios": "^1.5.1",
+    "classnames": "^2.3.2",
+    "moment": "^2.29.4",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hot-loader": "^4.13.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-transform-runtime": "^7.9.0",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-react": "^7.9.4",
-    "@hot-loader/react-dom": "^16.13.0",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.1.0",
+    "@babel/core": "^7.23.0",
+    "@babel/eslint-parser": "^7.22.15",
+    "@babel/plugin-transform-runtime": "^7.22.15",
+    "@babel/preset-env": "^7.22.20",
+    "@babel/preset-react": "^7.22.15",
+    "@hot-loader/react-dom": "^17.0.2",
+    "babel-loader": "^9.1.3",
     "babel-polyfill": "^6.26.0",
-    "css-loader": "^3.5.1",
-    "dotenv": "^8.2.0",
+    "css-loader": "^6.8.1",
+    "dotenv": "^16.3.1",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.1.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "html-loader": "^1.1.0",
-    "html-webpack-plugin": "^4.0.4",
+    "enzyme-adapter-react-16": "^1.15.7",
+    "eslint": "^8.51.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "html-loader": "^4.2.0",
+    "html-webpack-plugin": "^5.5.3",
     "ignore-loader": "^0.1.2",
-    "jest": "^25.2.7",
-    "less": "^3.11.1",
-    "less-loader": "^5.0.0",
-    "mini-css-extract-plugin": "^0.9.0",
-    "react-hot-loader": "^4.12.20",
-    "react-test-renderer": "^16.13.1",
-    "url-loader": "^4.0.0",
-    "webpack": "^4.42.1",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
+    "jest": "^29.7.0",
+    "less": "^4.2.0",
+    "less-loader": "^11.1.3",
+    "mini-css-extract-plugin": "^2.7.6",
+    "prettier": "^3.0.3",
+    "react-test-renderer": "^18.2.0",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   },
   "scripts": {
     "build": "webpack --mode production",
-    "start": "webpack-dev-server --hot --inline",
+    "start": "webpack-dev-server --hot",
     "lint": "eslint src",
+    "format:check": "prettier src --check",
+    "format:write": "prettier src --write",
     "test": "jest",
     "test:update": "jest --update-snapshot"
   },
